refactor(archtest): AFTERCOMMIT-HOOK-PURE-TRANSIENT-01 迁移到 module-path-agnostic [#1302]
规则逻辑从 _test.go 迁移到非 test 文件 aftercommit_pure_transient_invariants.go：
- 所有含 github.com/ghbvf/gocell 的 const 值（persistencePkgPath、
  afterCommitBannedReceivers map key、resolveBannedReceivers struct 元素）
  改为通过 PlatformModulePath 派生，不再有裸字面量
- 导出 CheckAfterCommitHookPureTransient(t, ConfigForExternalCell) []Diagnostic，
  cfg.BuildTags 透传到 RunTypedProduction TypedOpts.Tags（不再硬编 FlatNonDefaultTags）
- 所有 scan helper（scanRegisterAfterCommitHooks、scanHookBodyTransient、
  scanAfterCommitDrainCallers、scanRegisterAfterCommitEscapes 等）随 const 移入非 test 文件，
  _test.go fixture/RED 测试复用同一 helper，不留副本
- aftercommit_pure_transient_test.go 精简为 dogfood 单源 + fixture/RED 测试；
  // INVARIANT: 锚点在两文件各保留一份（per multi-file 规则约定）
- baseline 76 → 75 行（ratchet-down）

行为零变：现有 fixture/RED（A1A2/A3/Escapes/BlindSpots）+ ratchet 全部 GREEN。
